PythonBatchTools/VideoTextOverlay/requirements.txt
moviepy
numpy
decorator
imageio
imageio-ffmpeg
proglog
requests
tqdm
Pillow>=8.3.2